fastapi_mongo/requirements.txt
amqp
anyio
asgiref
azure-core
azure-storage-blob
bcrypt
billiard
celery
certifi
cffi
charset-normalizer
cryptography
fastapi
h11
idna
isodate
kombu
msrest
numpy
oauthlib
pandas
passlib
pathlib
prompt-toolkit
pycparser
pydantic
PyJWT
pymongo
python-dateutil
python-dotenv
python-multipart
pytz
requests
requests-oauthlib
six

typing_extensions
urllib3
uvicorn
vine
wcwidth
